requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
clip @ git+https://github.com/openai/CLIP.git@a1d071733d7111c9c014f024669f959182114e33
cmake==3.28.1
filelock==3.13.1
ftfy==6.1.3
idna==3.6
imageio==2.33.1
Jinja2==3.1.2
lazy_loader==0.3
lit==17.0.6
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
opencv-python==4.8.1.78
packaging==23.2
Pillow==10.1.0
regex==2023.10.3
requests==2.31.0
scikit-image==0.22.0
scipy==1.11.4
sympy==1.12
tifffile==2023.12.9
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tqdm==4.66.1
triton==2.0.0
typing_extensions==4.9.0
urllib3==2.1.0
wcwidth==0.2.12
